+++ server/readSqlServer.js
@@ -1,7 +1,7 @@
 const sql = require('mssql')
 
 const sqlConfig = {
-  domain: 'dsrm',
+  domain: process.env.WIN_DOMAIN,
   user: process.env.WIN_USER,
   password: process.env.WIN_PASSWORD,
   server: process.env.SQL_SERVER,
@@ -16,9 +16,9 @@ const sqlConfig = {
   }
 }
 
-async function readSqlServer (payload) {
+async function getItem (payload) {
   const {
-    item_no,
+    item_no
   } = payload
   try {
     await sql.connect(sqlConfig)
@@ -38,7 +38,7 @@ async function readSqlServer (payload) {
       }
     }
     // NAV
-    const queryNav = `SELECT * FROM [${process.env.NAV_DATABASE}].dbo.[DS Master - LIVE$Item] where [No_] = '${item_no}'`
+    const queryNav = `SELECT * FROM [${process.env.NAV_DATABASE}].dbo.[${process.env.NAV_ITEM_TABLE}] where [No_] = '${item_no}'`
     const res = await sql.query(queryNav)
 
     if (res.rowsAffected[0] > 0) {
@@ -51,7 +51,7 @@ async function readSqlServer (payload) {
     }
 
     // // Experlogix
-    const queryExp = `SELECT * FROM [${process.env.EXP_DATABASE}].dbo.[Experlogix_Items] where [PseudoItemNumber] = '${item_no}'`
+    const queryExp = `SELECT * FROM [${process.env.EXP_DATABASE}].dbo.[${process.env.EXP_ITEM_TABLE}] where [PseudoItemNumber] = '${item_no}'`
     const resExp = await sql.query(queryExp)
 
     if (resExp.rowsAffected[0] > 0) {
@@ -73,4 +73,62 @@ async function readSqlServer (payload) {
   }
 }
 
-module.exports = { readSqlServer }
+async function getOption (payload) {
+  const {
+    option_key
+  } = payload
+
+  const response = {
+    msg: 'No results',
+    results: {
+      exists: false,
+      list_price: null,
+      no: option_key,
+      family: null,
+      displayOption: null,
+      optionCode: null,
+      desc: null,
+      feature: null,
+      featureDesc: null,
+      xor: null,
+      requiredXOR: false,
+      componentItemNo: null,
+      qtyPer: null,
+      netAdder: false
+    }
+  }
+
+  try {
+    if (!option_key) throw Error('Option key (PseudoNo) is required')
+    await sql.connect(sqlConfig)
+
+    const query = `SELECT * FROM [${process.env.EXP_DATABASE}].dbo.[${process.env.EXP_OPTION_TABLE}] where [PseudoNo] = '${option_key}'`
+
+    const res = await sql.query(query)
+    if (res.rowsAffected[0] > 0) {
+      response.msg = 'Results from database'
+      const option = res.recordset[0]
+      response.results.exists = true
+      response.results.list_price = +option.List
+      response.results.family = option.FAMILY || null
+      response.results.displayOption = option.DisplayOption || null
+      response.results.optionCode = option.OptionCode || null
+      response.results.desc = option.OptionDescription || null
+      response.results.feature = option.Feature || null
+      response.results.featureDesc = option.FeatureDescription || null
+      response.results.xor = option.XOR || null
+      response.results.requiredXOR = option.requiredXOR || false
+      response.results.componentItemNo = option.ComponentItemPseudoNo || null
+      response.results.qtyPer = +option.QTYper || null
+      response.results.netAdder = option.NetAdder === 'True' ? true : false
+    }
+
+    sql.close()
+    return response
+  } catch (error) {
+    console.error()
+    return response
+  }
+}
+
+module.exports = { getItem, getOption }