+++ server/sql-query-routes.js
@@ -1,5 +1,6 @@
-const { readSqlServer } = require('./readSqlServer.js')
-const { updateItem } = require('./sqlUpdateItem')
+const { getItem, getOption } = require('./readSqlServer.js')
+const { updateItem } = require('./sqlUpdateItem.js')
+const { updateOption } = require('./sqlUpdateOption.js')
 
 async function routes (fastify, options) {
   // Test get
@@ -13,9 +14,9 @@ async function routes (fastify, options) {
     reply.send({ hello: 'From POST' })
   })
 
-  // Test READ
-  fastify.post('/read-sql', async (request, reply) => {
-    const resSql = await readSqlServer({
+  // Get Item from NAV && Experlogix
+  fastify.post('/getItem', async (request, reply) => {
+    const resSql = await getItem({
       item_no: request.body.item_no
     })
     console.log(resSql)
@@ -27,8 +28,22 @@ async function routes (fastify, options) {
     })
   })
 
+  // Get Option from NAV && Experlogix
+  fastify.post('/getOption', async (request, reply) => {
+    const resSql = await getOption({
+      option_key: request.body.option_key
+    })
+    console.log(resSql)
+    reply.send({
+      msg: 'Read SQL route',
+      body: {
+        ...resSql
+      }
+    })
+  })
+
   // Update Item
-  fastify.post('/update-item', async (request, reply) => {
+  fastify.post('/updateItem', async (request, reply) => {
     const resSql = await updateItem({
       item_no: request.body.item_no || null,
       app: request.body.app,
@@ -36,6 +51,20 @@ async function routes (fastify, options) {
       standard_cost: request.body.standard_cost || null,
       update_list_price: request.body.update_list_price || false,
       update_standard_cost: request.body.update_standard_cost || false
+    })
+    console.log(
+      resSql
+    )
+    reply.send({
+      body: resSql
+    })
+  })
+
+  // Update Option
+  fastify.post('/updateOption', async (request, reply) => {
+    const resSql = await updateOption({
+      key: request.body.key || null,
+      list_price: request.body.list_price || null
     })
     console.log(
       resSql