+++ ui-client/src/components/InfoCard.vue
@@ -14,7 +14,7 @@
           v-for="i in config[type].listItems"
           :key="i.prop"
           :label="i.label"
-          :value="itemInfo[i.prop]"
+          :value="itemInfo[i.prop] || '-'"
         />
       </ul>
     </template>
@@ -27,7 +27,7 @@ import ListItemKeyValueVue from './ListItemKeyValue.vue'
 // props
 defineProps({
   type: {
-    type: String
+    type: String // nav, experlogix, option
   },
   itemInfo: {
     type: Object
@@ -72,6 +72,64 @@ const config = {
       {
         label: 'List Price',
         prop: 'list_price_experlogix'
+      }
+    ]
+  },
+  option: {
+    exists: 'exists',
+    title: 'Experlogix Option',
+    listItems: [
+      {
+        label: 'List Price',
+        prop: 'list_price'
+      },
+      {
+        label: 'Key (PseudoNo)',
+        prop: 'no'
+      },
+      {
+        label: 'Family',
+        prop: 'family'
+      },
+      {
+        label: 'Display Option',
+        prop: 'displayOption'
+      },
+      {
+        label: 'Option Code',
+        prop: 'optionCode'
+      },
+      {
+        label: 'Description',
+        prop: 'desc'
+      },
+      {
+        label: 'Feature',
+        prop: 'feature'
+      },
+      {
+        label: 'Feature Description',
+        prop: 'featureDesc'
+      },
+      {
+        label: 'XOR',
+        prop: 'xor'
+      },
+      {
+        label: 'XOR Required',
+        prop: 'requiredXOR'
+      },
+      {
+        label: 'Component Item No',
+        prop: 'componentItemNo'
+      },
+      {
+        label: 'Qty Per',
+        prop: 'qtyPer'
+      },
+      {
+        label: 'Net Adder',
+        prop: 'netAdder'
       }
     ]
   }