+++ ui-client/src/components/ListItemKeyValue.vue
@@ -16,7 +16,7 @@ defineProps({
     required: true
   },
   value: {
-    type: [String, Number],
+    type: [String, Number, Boolean],
     required: true
   }
 })