+++ ui-client/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+# For Release
+- Handle letter case on item input [#1383](https://github.com/dascosales/dev-team/issues/1383)
+- Add Experlogix option update [#1384](https://github.com/dascosales/dev-team/issues/1384)
+
 # [0.2.0] 2022-05-16
 - Small refactors
 - Builds for Production